fix(model-datastructure): cache entries acquired through get PrefetchCache.getAll
Cache entries when using `getAll` like they are cached when using get `get`.
This guarantees that after accessing the values once no requests are required.
