+++ model-datastructure/src/commonMain/kotlin/org/modelix/model/lazy/CLVersion.kt
@@ -19,7 +19,6 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toInstant
-import org.modelix.model.IKeyListener
 import org.modelix.model.IKeyValueStore
 import org.modelix.model.IVersion
 import org.modelix.model.LinearHistory
@@ -401,58 +400,6 @@ private fun trackAccessedEntries(store: IKeyValueStore, body: (IDeserializingKey
     val objectStore = ObjectStoreCache(accessTrackingStore)
     body(objectStore)
     return accessTrackingStore.accessedEntries
-}
-
-private class AccessTrackingStore(val store: IKeyValueStore) : IKeyValueStore {
-    val accessedEntries: MutableMap<String, String?> = LinkedHashMap()
-
-    override fun newBulkQuery(deserializingCache: IDeserializingKeyValueStore, config: BulkQueryConfiguration): IBulkQuery {
-        return store.newBulkQuery(deserializingCache, config)
-    }
-
-    override fun get(key: String): String? {
-        val value = store.get(key)
-        accessedEntries.put(key, value)
-        return value
-    }
-
-    override fun getIfCached(key: String): String? {
-        val value = store.getIfCached(key)
-        if (value != null) {
-            accessedEntries[key] = value
-        }
-        return value
-    }
-
-    override fun put(key: String, value: String?) {
-        TODO("Not yet implemented")
-    }
-
-    override fun getAll(keys: Iterable<String>): Map<String, String?> {
-        val entries = store.getAll(keys)
-        accessedEntries.putAll(entries)
-        return entries
-    }
-
-    override fun putAll(entries: Map<String, String?>) {
-        TODO("Not yet implemented")
-    }
-
-    override fun prefetch(key: String) {
-        TODO("Not yet implemented")
-    }
-
-    override fun listen(key: String, listener: IKeyListener) {
-        TODO("Not yet implemented")
-    }
-
-    override fun removeListener(key: String, listener: IKeyListener) {
-        TODO("Not yet implemented")
-    }
-
-    override fun getPendingSize(): Int {
-        TODO("Not yet implemented")
-    }
 }
 
 fun CLVersion.runWrite(idGenerator: IIdGenerator, author: String?, body: (IWriteTransaction) -> Unit): CLVersion {
